requirements.txt
fvcore
xformers
segment-anything
ultralytics
typeguard
torchmetrics
blenderproc
opencv-python 
omegaconf
ruamel.yaml
hydra-colorlog
hydra-core
triton
gdown
pandas
imageio
rich
pyrender
pytorch-lightning
pycocotools 
distinctipy
timm
gorilla-core
trimesh
gpustat
imgaug
einops
pyrealsense2
keyboard